fixed conflicts and added new gapi stuff
